requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3
anthropic==0.34.2
python-dotenv==1.0.1